Implemented a theme system that allows for custom HTML/CSS/JS content.
Themes are directories that contain `static` and `views` directories.
The `static` directory is mounted at `/theme/`, and any templates in the
`views` directory override Aestimia's standard templates. The primary
one to override is `layout.html`, which should extend `base-layout.html`.
This file can define its own `theme_head` and `theme_footer` blocks;
we can always add more blocks to `base-layout.html` if we need to make
it possible for themes to override other parts of pages.

I haven't yet documented this fully in README.md, because I wanted to
try out the system for a bit before investing too much into it, but
it seems decent and extensible enough.