Mentors must now assign submissions to themselves before assessing.
They have 15 minutes to assess a submission by default, at which point
someone else can assess it.

Note that this is not intended as a security mechanism, as it's technically
still possible for a properly-formed POST request to add a review to
the assessment; this is simply done as an aid to avoid race conditions.